FileAPI/new-Blob.html

<!doctype html>
<html>
 <head>
  <title>Blob constructor</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <iframe style=display:none></iframe><!-- used by "platform object that supports indexed properties" test -->
  <script>
    test(function(){
        var b = new Blob();
        assert_true(b instanceof Blob, 'b instanceof Blob');
        assert_equals(String(b), '[object Blob]', 'String(b)');
        assert_equals(b.size, 0, 'b.size');
        assert_equals(b.type, '', 'b.type');
    }, document.title+', no args');

    async_test(function(){
        var b1 = new Blob(['\r\r\n\n'], {endings:'native'});
        assert_equals(b1.size, 4);
        var reader = new FileReader();
        reader.onload = this.step_func(function(e) {
            assert_equals(reader.result, '\r\r\n\n');
            this.done();
        });
        reader.onerror = this.step_func(function(e) {
            assert_unreached('got error event on FileReader');
        });
        reader.readAsText(b1);
    }, document.title+', non-support for obsolete "endings" thing');

    test(function(){
        assert_false('toNativeLineEndings' in window);
    }, document.title+', non-support for obsolete "toNativeLineEndings" thing');

    test(function() {
        assert_throws(new TypeError(), function() { new Blob("foo"); });
        assert_throws(new TypeError(), function() { new Blob(document.createElement("div")); });
        assert_throws(new TypeError(), function() { new Blob([{toString:null, valueOf:null}]); });
    }, document.title+', wrong first arg');

    test(function() {
        var b = new Blob({"0": "foo", "length":1});
        assert_equals(b.size, 3);
    }, document.title+', object with properties "0" and "length"');

    test(function() {
        var b_window = new Blob(window);
        assert_equals(b_window.size, '[object Window]'.length, 'window');
        window[0].toString = function() { return 'foo'; };
        var b_window_2 = new Blob(window);
        assert_equals(b_window_2.size, 3, 'window');
        var select = document.createElement('select');
        select.appendChild(document.createElement('option'));
        var b_select = new Blob(select);
        assert_equals(b_select.size, '[object HTMLOptionElement]'.length, 'select');
    }, document.title+', platform object that supports indexed properties');

    test(function() {
        var list = document.body.children;
        Object.defineProperty(list, "length", {get: function() { throw {name:7}; }});
        assert_throws({name:7}, function(){ new Blob(list); });
    }, document.title+', platform object that supports indexed properties, overwriting "length"');

    var t_ports = async_test(document.title+', MessagePort[] (platform array object)');
    t_ports.step(function() {
        var channel = new MessageChannel();
        channel.port2.onmessage = this.step_func(function(e) {
            var b_ports = new Blob(e.ports);
            assert_equals(b_ports.size, '[object MessagePort]'.length);
            this.done();
        });
        var channel2 = new MessageChannel();
        channel.port1.postMessage('', [channel2.port1]);
    });

    test(function() {
        var elm = document.createElement('div');
        elm.setAttribute('foo', 'bar');
        var b_attrs = new Blob(elm.attributes);
        assert_equals(b_attrs.size, '[object Attr]'.length);
    }, document.title+', Attr[] (platform array object)');

    var type_tests = [
    // blobParts, type, expected type
    [[], 'a', 'a'],
    [[], 'A', 'a'],
    [[], '\u00E5', ''],
    [[], '\uD801\uDC7E', ''], // U+1047E
    [[], ' image/gif ', ' image/gif '],
    [[], '\timage/gif\t', ''],
    [[], 'image/gif;\u007f', ''],
    [[], '\u0130mage/gif', ''], // uppercase i with dot
    [[], '\u0131mage/gif', ''], // lowercase dotless i
    [[], 'image/gif\u0000', ''],
    // check that type isn't changed based on sniffing
    [[0x3C, 0x48, 0x54, 0x4D, 0x4C, 0x3E], 'unknown/unknown', 'unknown/unknown'], // "<HTML>"
    [[0x00, 0xFF], 'text/plain', 'text/plain'],
    [[0x47, 0x49, 0x46, 0x38, 0x39, 0x61], 'image/png', 'image/png'], // "GIF89a"
    ];

    type_tests.forEach(function(t) {
        test(function() {
            var arr = new Uint8Array([t[0]]).buffer;
            var b = new Blob([arr], {type:t[1]});
            assert_equals(b.type, t[2]);
        }, document.title+', type: '+t[1]);
    });

    test(function() {
        var b1 = new Blob(['foo']);
        var b2 = new Blob([b1, b1]);
        assert_equals(b2.size, 6);
    }, document.title+', Blob in first argument');

    test(function() {
        var view = new Uint8Array([0, 255, 0]);
        var b1 = new Blob([view.buffer, view.buffer]);
        assert_equals(b1.size, 6);
    }, document.title+', ArrayBuffer in first argument');

    test(function() {
        var view = new Uint8Array([0, 255, 0, 4]);
        var b1 = new Blob([view, view]);
        assert_equals(b1.size, 8);
        var view1 = new Uint16Array(view.buffer, 2);
        var b2 = new Blob([view1, view.buffer, view1]);
        assert_equals(b2.size, 2 + 4 + 2);
    }, document.title+', ArrayBufferView in first argument');

    test(function() {
        var view = new Uint8Array([0]);
        var b1 = new Blob(['fo']);
        var b2 = new Blob([view.buffer, b1, "foo"]);
        assert_equals(b2.size, 6);
    }, document.title+', mixed types in first argument');

    test(function() {
        var b1 = new Blob([null, undefined]);
        assert_equals(b1.size, "nullundefined".length);
    }, document.title+', null/undefined in first argument');

    var t_slice = async_test(document.title+', slice');
    t_slice.step(function(){
        var view = new Uint8Array([0, 255, 0]);
        var b1 = new Blob(['abcd']);
        var b2 = new Blob([view.buffer, b1, 'efgh', 'ijklmnopqrstuvwxyz']);
        var b3 = b2.slice(1, 4);
        b3.expectedText = '\uFFFD\u0000a';
        b3.expectedSize = 3;
        b3.msg = 'b3';
        var b4 = b2.slice(4, 8);
        b4.expectedText = 'bcde';
        b4.expectedSize = 4;
        b4.msg = 'b4';
        var b5 = b2.slice(8, 12);
        b5.expectedText = 'fghi';
        b5.expectedSize = 4;
        b5.msg = 'b5';
        var b6 = b2.slice(1, 12);
        b6.expectedText = '\uFFFD\u0000abcdefghi';
        b6.expectedSize = 11;
        b6.msg = 'b6';
        [b3, b4, b5, b6].forEach(this.step_func(function(b) {
            assert_equals(b.size, b.expectedSize, b.msg);
            var reader = new FileReader();
            reader.onload = this.step_func(function(e) {
                assert_equals(reader.result, b.expectedText, b.msg);
                if (b === b6)
                    this.done();
            });
            reader.readAsText(b, 'utf-8');
        }));
    });
  </script>
 </body>
</html>
